requirements.txt
pandas
numpy
scikit-learn
meteostat
requests
python-dateutil
joblib
matplotlib
